add sound-card-select.sh
This script uses yad to ask the user to select the
sound card and buffer size and then starts jamulus.
In case the user wants to use different sound cards
for input and output, alsa_in is started.
To offer control over the audio volume, jack_mixer is
also started and all sound sources and sinks are
connected.

Signed-off-by: Martin Kaistra <admin@djfun.de>